+++ CHANGELOG.md
@@ -34,3 +34,38 @@ Sun 13 Jun 2021
 
 - Added the possibility to install some GNOME applications
 - Added the possibility to install some KDE applications
+
+Version 3.2 <br>
+Sat 19 Jun 2021
+
+- Added package engrampa to MATE
+- Replaced engrampa with xarchiver package for Xfce
+- Added LXDE desktop
+- Added file-roller to GNOME, Budgie and Cinnamon
+- Added package ffmpegthumbs to KDE Plasma to display video thumbnails in Dolphin
+
+Version 3.3 <br>
+Fri 02 Jul 2021
+
+- Replaced sddm with lightdm for LXQt
+- Added qterminal to LXQt
+- Some improvements in text output
+
+Version 4.0 <br>
+Sat 31 Jul 2021
+
+- Added missing package kde-gtk-config5 to KDE Plasma
+- Added possibility to install KDE Connect
+- Added FVWM3 window manager
+- Added missing Audio support install option
+- Added install option for terminal emulators
+- Added package bash-completion to recommended packages
+- Added package neofetch to xorg installation
+
+Version 4.1 <br>
+Fri 06 Aug 2021
+
+- Added mate-polkit and mate-power-manager to MATE setup
+- Added xarchiver to Enlightenment and LXDE setup
+- Added German language pack to Firefox and Firefox ESR setup
+- Added German and American English language pack to LibreOffice setup